get the delete button to work
