+++ app/src/main/java/com/dil/notifier/MainActivity.kt
@@ -151,41 +151,27 @@ class MainActivity : AppCompatActivity() {
      * When the delete button is clicked, delete the notification settinsg for the app name that is currently typed
      */
     fun deleteNotification(name: String) {
-        //Delete as long as user inputted a name
-        if (name != "") {
-            val builder: AlertDialog.Builder = AlertDialog.Builder(this)
-            builder
-                .setMessage("Delete your custom notification for " + name + "?")
-                .setTitle("Are you sure?")
-                .setPositiveButton("Delete") { dialog, which ->
-                    //Get list of notification channels
-                    var list1: List<NotificationChannel> = ArrayList()
-                    val notificationManager =
-                        getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-                    list1 = notificationManager.notificationChannels
+        //Get list of notification channels
+        var list1: List<NotificationChannel> = ArrayList()
+        val notificationManager =
+            getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+        list1 = notificationManager.notificationChannels
 
-                    for (notif in list1) {
-                        val notifName = notif.id
-                        if (notifName.startsWith(name + "ScreenOn")) {
-                            notificationManager.deleteNotificationChannel(notifName)
-                        }
-                        if (notifName.startsWith(name + "ScreenOff")) {
-                            notificationManager.deleteNotificationChannel(notifName)
-                        }
-                    }
-
-                    //Tell user it's been deleted
-                    Toast.makeText(this, name + " notification deleted", Toast.LENGTH_SHORT).show()
-
-                    //Update list of notification channels using function defined earlier
-                    getList()
-                }
-                .setNegativeButton("Cancel") { dialog, which ->
-                    // Do nothing
-                }
-            val dialog: AlertDialog = builder.create()
-            dialog.show()
+        for (notif in list1) {
+            val notifName = notif.id
+            if (notifName.startsWith(name + "ScreenOn")) {
+                notificationManager.deleteNotificationChannel(notifName)
+            }
+            if (notifName.startsWith(name + "ScreenOff")) {
+                notificationManager.deleteNotificationChannel(notifName)
+            }
         }
+
+        //Tell user it's been deleted
+        Toast.makeText(this, name + " notification deleted", Toast.LENGTH_SHORT).show()
+
+        //Update list of notification channels using function defined earlier
+        getList()
     }
 
     //When app is created
@@ -235,7 +221,7 @@ class MainActivity : AppCompatActivity() {
         //Update list of notification channels using function defined earlier
         getList()
 
-        var notifications = listOf(NotificationData("1", "One") { deleteNotification("test2") })
+        var notifications = mutableListOf(NotificationData("1", "One") { deleteNotification("test2") })
         val notiList: RecyclerView = findViewById(R.id.notiList)
         val layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
         val myAdapter = NotificationListAdapter(notifications)
